harmless clang warnings

+++ vcg/complex/algorithms/point_outlier.h
@@ -58,7 +58,7 @@ class OutlierRemoval
       typename MeshType::template PerVertexAttributeHandle<ScalarType> plof =         tri::Allocator<MeshType>:: template GetPerVertexAttribute<ScalarType>(mesh, std::string("plof"));
 
 #pragma omp parallel for schedule(dynamic, 10)
-      for (int i = 0; i < mesh.vert.size(); i++)
+      for (size_t i = 0; i < mesh.vert.size(); i++)
       {
         PriorityQueue queue;
         kdTree.doQueryK(mesh.vert[i].cP(), kNearest, queue);
@@ -71,7 +71,7 @@ class OutlierRemoval
 
       float mean = 0;
 #pragma omp parallel for reduction(+: mean) schedule(dynamic, 10)
-      for (int i = 0; i < mesh.vert.size(); i++)
+      for (size_t i = 0; i < mesh.vert.size(); i++)
       {
         PriorityQueue queue;
         kdTree.doQueryK(mesh.vert[i].cP(), kNearest, queue);
@@ -87,7 +87,7 @@ class OutlierRemoval
       mean = sqrt(mean);
 
 #pragma omp parallel for schedule(dynamic, 10)
-      for (int i = 0; i < mesh.vert.size(); i++)
+      for (size_t i = 0; i < mesh.vert.size(); i++)
       {
         ScalarType value = plof[i] / (mean * sqrt(2.0f));
         double dem = 1.0 + 0.278393 * value;
